Add FramedRing datastructure, more sound direct primitives, and packet
Lot of stuff is going on here (a lot has changed since I last
committed!)... This is primarily trying to clean up cursor.rs madness to
some degree (and on the other hand avoid it).

cursor.rs has a lot of problems - it does too much and at the same time
not enough. Multibytes & the Cursor primative are bad designs and
inflexible. direct.rs takes on some blockrouter specific changes to
buffers which we need for optimizations, while buffer.rs begins to try
to be a more modern multi-'list' container. The idea is that FramedRing
will in the end avoid needing to allocate memory for each packet
recieved (and many times over when stuff like the compressor or framer
needed to create new VecDeques.

buffer.rs implements FramedRing, a ring buffer (like VecDeque) which has
multiple 'frames' - only the head frame may append while other frames
may be read from at any time (and are cleaned up as they go out of
scope). This is sort of the usage pattern I would expect for a packet
processor - take previous input (such as a packet being a collection of
direct bufs) and 'construct' a new packet within either the same or
another FramedRing - all without needing a guarenteed memory allocation
each time we create a new one of these 'frames'.

What could possibly go wrong?